src/test/java/com/example/springdatajpa_datajpa/dao/AuthorDaoJdbcTemplateTest.java
package com.example.springdatajpa_datajpa.dao;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.context.annotation.ComponentScan;
import org.springframework.data.domain.PageRequest;
import org.springframework.jdbc.core.JdbcTemplate;

@DataJpaTest
@ComponentScan(basePackages = "com.example.springdatajpa_datajpa.dao")
@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
public class AuthorDaoJdbcTemplateTest {

    @Autowired
    JdbcTemplate jdbcTemplate;

    AuthorDao authorDao;

    @BeforeEach
    void setUp() {
        authorDao = new AuthorDaoJdbcTemplate(jdbcTemplate);
    }

    @Test
    void testDeleteAuthorById() {

    }

    @Test
    void testFindAllAuthorsByLastName() {
        var authors = authorDao.findAllAuthorsByLastName("Smith", PageRequest.of(0,10));
        assertThat(authors).isNotNull();
        assertThat(authors.size()).isEqualTo(10);
    }

    @Test
    void testFindAuthorByName() {

    }

    @Test
    void testGetById() {

    }

    @Test
    void testSaveNewAuthor() {

    }

    @Test
    void testUpdateAuthor() {

    }
}
